+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [3.14.00] - (11.12.2021)
+## Changes
+- added read sound mode from device
+- update config.schema
+- stability and performance improvements
+
 ## [3.13.03] - (11.12.2021)
 ## Changes
 - use event emmiter for state changes